+++ uniquify/src/CSS/Compare.module.css
@@ -1,6 +1,6 @@
-body {
+/* body {
   background-color: #1a1a1a;
-}
+} */
 
 input {
   border-radius: 10px;
